requirements.txt
earthengine-api
geemap
click
Pillow
numpy
tifffile
tqdm
torch
torchvision
